% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPIChelperfunctions.R
\name{convert_to_SCE}
\alias{convert_to_SCE}
\title{Convert Sparse Matrix to SingleCellExperiment object}
\usage{
convert_to_SCE(sparse_data)
}
\arguments{
\item{sparse_data}{A sparse matrix containing count data, where rows are
peaks and columns represent cells.}
}
\value{
A SingleCellExperiment (SCE) object with the sparse matrix stored
in the "counts" assay.
}
\description{
This function converts a sparse matrix into a SingleCellExperiment (SCE)
object.
}
