% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPIChelperfunctions.R
\name{getCounts}
\alias{getCounts}
\title{Get counts from Single Cell Experiment object}
\usage{
getCounts(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{
counts matrix
}
\description{
Get counts matrix from a SingleCellExperiment object. If counts is
missing a warning is issued and the first assay is returned.
}
