% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICsimulate.R
\name{getLNormFactors}
\alias{getLNormFactors}
\title{Get accessibility factors}
\usage{
getLNormFactors(n.facs, sel.prob, neg.prob, fac.loc, fac.scale)
}
\arguments{
\item{n.facs}{Number of factors to generate.}

\item{sel.prob}{Probability that a factor will be selected to be different
from 1.}

\item{neg.prob}{Probability that a selected factor is less than one.}

\item{fac.loc}{Location parameter for the log-normal distribution.}

\item{fac.scale}{Scale factor for the log-normal distribution.}
}
\value{
Vector containing generated factors.
}
\description{
Randomly generate multiplication factors from a log-normal distribution.
}
\keyword{internal}
