% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPIChelperfunctions.R
\name{newsimPICcount}
\alias{newsimPICcount}
\title{newsimPICcount}
\usage{
newsimPICcount(...)
}
\arguments{
\item{...}{Variables to set newsimPICcount object parameters.}
}
\value{
new object from class simPICcount
}
\description{
create a newsimPICcount object to store parameters
this function creates the object variable which is passed in all
functions.
}
\examples{
object <- newsimPICcount()

}
