% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICcompare.R
\name{plot_theme}
\alias{plot_theme}
\title{Custom theme for ggplot2}
\usage{
plot_theme()
}
\value{
A ggplot2 theme object with predefined settings.
}
\description{
This function defines a custom theme for ggplot2 to ensure consistent
visual appearance across multiple plots.
}
