% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPIChelperfunctions.R
\name{setsimPICparameters}
\alias{setsimPICparameters}
\title{Set simPIC parameters}
\usage{
setsimPICparameters(object, update = NULL, ...)
}
\arguments{
\item{object}{input simPICcount object}

\item{update}{new parameters}

\item{...}{set new parameters for simPICcount object}
}
\value{
simPICcount object with updated parameters
}
\description{
Set input parameters of the simPICcount object
}
\examples{
object <- newsimPICcount()
object <- setsimPICparameters(object, nCells = 200, nPeaks = 500)

}
