% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICcompare.R
\name{simPICcompare}
\alias{simPICcompare}
\title{Compare SingleCellExperiment objects}
\usage{
simPICcompare(
  sces,
  point.size = 0.2,
  point.alpha = 0.1,
  fits = TRUE,
  colours = NULL
)
}
\arguments{
\item{sces}{named list of SingleCellExperiment objects to combine and
compare.}

\item{point.size}{size of points in scatter plots.}

\item{point.alpha}{opacity of points in scatter plots.}

\item{fits}{whether to include fits in scatter plots.}

\item{colours}{vector of colours to use for each dataset.}
}
\value{
List containing the combined datasets and plots.
}
\description{
Combine data from several SingleCellExperiment objects and produce
some basic plots comparing them.
}
\details{
The returned list has three items:
The first dataset in \code{sces} is treated as the reference dataset when
computing Kolmogorov-Smirnov (KS) summaries.

\describe{
   \item{\code{RowData}}{Combined row data from the provided
   SingleCellExperiments.}
   \item{\code{ColData}}{Combined column data from the provided
   SingleCellExperiments.}
   \item{\code{KS}}{Data frame summarizing KS statistics for peak means,
   library sizes, and cell sparsity, comparing each dataset to the first
   dataset in \code{sces}.}
   \item{\code{Plots}}{Comparison plots
       \describe{
           \item{\code{Means}}{Boxplot of mean distribution.}
           \item{\code{Variances}}{Boxplot of variance distribution.}
           \item{\code{MeanVar}}{Scatter plot with fitted lines showing
           the mean-variance relationship.}
           \item{\code{LibrarySizes}}{Boxplot of the library size
           distribution.}
           \item{\code{ZerosPeak}}{Boxplot of the percentage of each peak
           that is zero.}
           \item{\code{ZerosCell}}{Boxplot of the percentage of each cell
          that is zero.}
           \item{\code{MeanZeros}}{Scatter plot with fitted lines showing
           the mean-zeros relationship.}
   }
 }
}

The plots returned by this function are created using
\code{\link[ggplot2]{ggplot}} and are only a sample of the kind of plots
you might like to consider. The data used to create these plots is also
returned and should be in the correct format to allow you to create
further plots using \code{\link[ggplot2]{ggplot}}.
}
\examples{
sim1 <- simPICsimulate(
    nPeaks = 1000, nCells = 500,
    pm.distr = "weibull", seed = 7856
)
sim2 <- simPICsimulate(
    nPeaks = 1000, nCells = 500,
    pm.distr = "gamma", seed = 4234
)
comparison <- simPICcompare(list(weibull = sim1, gamma = sim2))
names(comparison)
names(comparison$Plots)
}
