% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICestimate.R
\name{simPICEstBCV}
\alias{simPICEstBCV}
\title{Estimate simPIC Biological Coefficient of Variation parameters}
\usage{
simPICEstBCV(counts, object, verbose)
}
\arguments{
\item{counts}{counts matrix to estimate parameters from.}

\item{object}{simPICcount object to store estimated values in.}

\item{verbose}{Logical. Whether to print progress messages.}
}
\value{
simPICcount object with estimated values.
}
\description{
Parameters are estimated using the \code{\link[edgeR]{estimateDisp}} function
in the \code{edgeR} package.
}
\details{
The \code{\link[edgeR]{estimateDisp}} function is used to estimate the common
dispersion and prior degrees of freedom. See
\code{\link[edgeR]{estimateDisp}} for details. When estimating parameters on
simulated data we found a broadly linear relationship between the true
underlying common dispersion and the \code{edgeR} estimate, therefore we
apply a small correction, \code{disp = -0.3 + 0.15 * edgeR.disp}.
}
