% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICestimate.R
\name{simPICestimatePeakMean}
\alias{simPICestimatePeakMean}
\title{Estimate simPIC peak means}
\usage{
simPICestimatePeakMean(norm.counts, object, pm.distr, verbose)
}
\arguments{
\item{norm.counts}{Library-size normalized count matrix.}

\item{object}{simPICcount object to store estimated values.}

\item{pm.distr}{distribution parameter for peak means.}

\item{verbose}{Logical. Whether to print progress messages.}
}
\value{
simPICcount object containing all estimated parameters
}
\description{
Estimate peak mean parameters for simPIC simulation
}
\details{
Parameters for gamma distribution are estimated by fitting the mean
normalized counts using \code{\link[fitdistrplus]{fitdist}}.
All the fitting methods are tried and the fit with the best Cramer-von
Mises statistic is selected.
}
