% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICestimate.R
\name{simPICestimateSparsity}
\alias{simPICestimateSparsity}
\title{Estimate sparsity.}
\usage{
simPICestimateSparsity(norm.counts, object, verbose)
}
\arguments{
\item{norm.counts}{A normalized count matrix to estimate parameters from.}

\item{object}{simPICcount object to store estimated parameters.}

\item{verbose}{logical. To print messages or not.}
}
\value{
simPICcount object with updated sparsity parameter.
}
\description{
This function estimates the sparsity of cells based on a normalized counts 
matrix and updates the parameters of a simPIC object accordingly.
}
