% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPIChelperfunctions.R
\name{simPICget}
\alias{simPICget}
\title{Get a single simPICcount parameter}
\usage{
simPICget(object, name)
}
\arguments{
\item{object}{input simPICcount object.}

\item{name}{name of the parameter.}
}
\value{
Value of the input parameter.
}
\description{
Get the value of a single variable from input simPICcount object.
}
\examples{
object <- newsimPICcount()
nPeaks <- simPICget(object, "nPeaks")

}
