% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPIChelperfunctions.R
\name{simPICgetparameters}
\alias{simPICgetparameters}
\title{Get parameters}
\usage{
simPICgetparameters(object, names)
}
\arguments{
\item{object}{input object to get values from.}

\item{names}{vector of names of the parameters to get.}
}
\value{
List with the values of the selected parameters.
}
\description{
Get multiple parameter values from a simPIC object.
}
\examples{
object <- newsimPICcount()
simPICgetparameters(object, c("nPeaks", "nCells", "peak.mean.shape"))
}
