% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICsimulate.R
\name{simPICsimBatchCellMeans}
\alias{simPICsimBatchCellMeans}
\title{Simulate batch means}
\usage{
simPICsimBatchCellMeans(object, sim)
}
\arguments{
\item{object}{simPICcount object with simulation parameters.}

\item{sim}{SingleCellExperiment to add batch means to.}
}
\value{
SingleCellExperiment with simulated batch means.
}
\description{
Simulate a mean for each peak in each cell incorporating batch effect
factors.
}
\keyword{internal}
