% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICsimulate.R
\name{simPICsimBatchEffects}
\alias{simPICsimBatchEffects}
\title{Simulate batch effects}
\usage{
simPICsimBatchEffects(object, sim)
}
\arguments{
\item{object}{simPICcount object with simulation parameters.}

\item{sim}{SingleCellExperiment to add batch effects to.}
}
\value{
SingleCellExperiment with simulated batch effects.
}
\description{
Simulate batch effects. Batch effect factors for each batch are produced
using \code{\link{getLNormFactors}} and these are added along with updated
means for each batch.
}
\keyword{internal}
