% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICsimulate.R
\name{simPICsimCellMeans}
\alias{simPICsimCellMeans}
\alias{simPICsimSingleCellMeans}
\alias{simPICsimulateGroupCellMeans}
\title{Simulate cell means}
\usage{
simPICsimSingleCellMeans(object, sim)

simPICsimulateGroupCellMeans(object, sim)
}
\arguments{
\item{object}{simPIC object with simulation parameters.}

\item{sim}{SingleCellExperiment to add cell means to.}
}
\value{
SingleCellExperiment with added cell means.
}
\description{
Simulate a peak by cell matrix given the mean accessibility for each peak in
each cell. Cells start with the mean accessibility for the group they belong
to (when simulating groups). The selected means are adjusted for each cell's
expected library size.
}
