% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICsimulate.R
\name{simPICsimulateBCVMeans}
\alias{simPICsimulateBCVMeans}
\title{Simulate BCV means}
\usage{
simPICsimulateBCVMeans(object, sim)
}
\arguments{
\item{object}{simPICcount object with simulation parameters.}

\item{sim}{SingleCellExperiment to add BCV means to.}
}
\value{
SingleCellExperiment with simulated BCV means.
}
\description{
Simulate means for each peak in each cell that are adjusted to follow a
mean-variance trend using Biological Coefficient of Variation taken from
and inverse gamma distribution.
}
