% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICsimulate.R
\name{simPICsimulatemultiDA}
\alias{simPICsimulatemultiDA}
\title{Simulate group differential accessibility}
\usage{
simPICsimulatemultiDA(object, sim)
}
\arguments{
\item{object}{simPICcount object with simulation parameters.}

\item{sim}{SingleCellExperiment to add differential accessibility to.}
}
\value{
SingleCellExperiment with simulated differential accessibility.
}
\description{
Simulate differential accessibility. Differential accessibility factors for each
group are produced using \code{\link{getLNormFactors}} and these are added
along with updated means for each group. For paths care is taken to make sure
they are simulated in the correct order.
}
