% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICsimulate.R
\name{simPICsimulateTrueCountsGroups}
\alias{simPICsimulateTrueCountsGroups}
\title{Simulate true counts groups.}
\usage{
simPICsimulateTrueCountsGroups(object, sim)
}
\arguments{
\item{object}{simPICcount object with simulation parameters.}

\item{sim}{SingleCellExperiment object containing simulation parameters.}
}
\value{
SingleCellExperiment object with simulated true counts.
}
\description{
Counts are simulated from a poisson distribution where each peak has a
mean, expected library size and proportion of accessible chromatin.
}
