% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICestimate.R
\name{selectFit}
\alias{selectFit}
\title{Select fit}
\usage{
selectFit(data, distr, verbose = TRUE)
}
\arguments{
\item{data}{The data to fit}

\item{distr}{Name of the distribution to fit}

\item{verbose}{logical. To print messages or not}
}
\value{
The selected fit object
}
\description{
Trying a variety of fitting methods and selecting the best one
}
\details{
The distribution is fitted to the data using each of the
\code{\link[fitdistrplus]{fitdist}} fitting methods. The fit with the
smallest Cramer-von Mises statistic is selected.
}
