% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPIC-package.R
\docType{package}
\name{simPIC-package}
\alias{simPIC}
\alias{simPIC-package}
\title{simPIC: Flexible simulation of paired-insertion counts for single-cell ATAC-sequencing data}
\description{
simPIC is a package for simulating single-cell ATAC-seq count data. It provides a user-friendly, well documented interface for data simulation. Functions are provided for parameter estimation, realistic scATAC-seq data simulation, and comparing real and simulated datasets.

\itemize{
\item count class (\code{\link{newsimPICcount}})
\item estimate (\code{\link{simPICestimate}})
\item simulate (\code{\link{simPICsimulate}})
\item plots (\code{\link{simPICcompare}})
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sagrikachugh/simPIC}
  \item Report bugs at \url{https://github.com/sagrikachugh/simPIC/issues}
}

}
\author{
\strong{Maintainer}: Sagrika Chugh \email{sagrika.chugh@gmail.com} (\href{https://orcid.org/0000-0002-8050-5214}{ORCID})

Authors:
\itemize{
  \item Heejung Shim
  \item Davis McCarthy
}

}
\keyword{internal}
