% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPIC-package.R
\docType{package}
\name{simPIC-package}
\alias{simPIC}
\alias{simPIC-package}
\title{simPIC: A uniform quantification based method for simulating Paired-Insertion count matrices for single-cell ATAC sequencing data}
\description{
simPIC is a package for simulating single-cell ATAC-seq count data. It provides a user-friendly, well documented interface for data simulation. Functions are proided for parameter estimation, realistic scATAC-seq data simulation, and comparing real and simulated datasets.

\itemize{
\item count class (\code{\link{newsimPICcount}})
\item estimate (\code{\link{simPICestimate}})
\item simulate (\code{\link{simPICsimulate}})
\item plots (\code{\link{simPICcompare}})
}
}
\author{
\strong{Maintainer}: Sagrika Chugh \email{sagrika.chugh@gmail.com} (\href{https://orcid.org/0000-0002-8050-5214}{ORCID})

}
\keyword{internal}
