% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICestimate.R
\name{simPICestimateSparsity}
\alias{simPICestimateSparsity}
\title{Estimate sparsity}
\usage{
simPICestimateSparsity(norm.counts, object, verbose)
}
\arguments{
\item{norm.counts}{A normalized count matrix to estimate parameters from.}

\item{object}{simPICcount object to store estimated parameters.}

\item{verbose}{Logical. Whether to print progress messages.}
}
\value{
simPICcount object with updated sparsity parameter.
}
\description{
This function estimates cell sparsity from a normalized count matrix and
updates the parameters of a simPIC object accordingly.
}
