% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICestimate.R
\name{simPICestimateSparsity}
\alias{simPICestimateSparsity}
\title{Estimate simPIC peak sparsity.}
\usage{
simPICestimateSparsity(norm.counts, object, verbose)
}
\arguments{
\item{norm.counts}{A sparse count matrix to estimate parameters from.}

\item{object}{simPICcount object to store estimated parameters.}

\item{verbose}{logical. To print messages or not.}
}
\value{
simPICcount object with updated non-zero cell proportion
parameter.
}
\description{
Extract the accessibility proportion (sparsity) of each cell among all
peaksvfrom the input count matrix.
}
\details{
Vector of non-zero cell proportions of peaks is calculated by
dividing the number of non-zero entries over the number of all cells
for each peak.
}
