% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICsimulate.R
\name{simPICsimulateLibSize}
\alias{simPICsimulateLibSize}
\title{Simulate simPIC library sizes}
\usage{
simPICsimulateLibSize(object, sim, verbose)
}
\arguments{
\item{object}{simPICcount object with simulation parameters.}

\item{sim}{SingleCellExperiment object containing simulation parameters.}

\item{verbose}{logical. To print progress messages}
}
\value{
SingleCellExperiment object with simulated library sizes
}
\description{
Generate library sizes for cells in simPIC simulation based on the
estimated values of mus and sigmas.
}
