% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPICsimulate.R
\name{simPICsimulatePeakMean}
\alias{simPICsimulatePeakMean}
\title{Simulate simPIC peak means.}
\usage{
simPICsimulatePeakMean(object, sim, pm.distr, verbose)
}
\arguments{
\item{object}{simPICcount object with simulation parameters.}

\item{sim}{SingleCellExperiment object containing simulation parameters.}

\item{pm.distr}{distribution parameter for peak means.
Available distributions: gamma, weibull, lngamma, pareto.
Default is weibull.}

\item{verbose}{logical. Whether to print progress messages.}
}
\value{
SingleCellExperiment object with simulated peak means.
}
\description{
Generate peak means for cells in simPIC simulation based on the estimated
values of shape and rate parameters.
}
